requirements.txt
# Core Django & Extensions
Django==5.0.14
django-crispy-forms>=2.1.0
django-import-export>=3.3.7
django-taggit>=5.0.1
djangorestframework>=3.14.0
django-unfold>=0.22.0
django-ckeditor-5>=0.2.12
django-js-asset>=3.1.2

# Database
dj-database-url>=2.1.0
psycopg2-binary>=2.9.6

# Image & File Processing
Pillow>=10.2.0
pyuploadcare>=4.1.0

# Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Environment & Configuration
python-decouple>=3.8
python-dotenv>=1.0.1

# CORS & Security
django-cors-headers>=4.3.1
sentry-sdk>=1.40.0

# HTTP/API
requests>=2.31.0

# Google API related
google-auth>=2.28.1
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0

# File handling
openpyxl>=3.1.2
PyYAML>=6.0.1

# HTML/Text Processing
html5lib>=1.1
shortuuid>=1.0.13

# Background Tasks
django-q2>=1.8.0

# Testing
coverage>=7.4.1
pytest>=8.0.0
pytest-django>=4.7.0
